reorganization and toolbar ui complete
