+++ components/board/Square.js
